Fix null exception
We initialize the entire array with defaults, but not every item gets filled in during PML scan. Any direct query of that array may have null frames for example. Could go preinit with Array.Empty but it's better to just force callers to think about whether or not there is a record there.