Restore CaptureTime
Don't force callers to convert to DateTime and back again, if what they really want is the underlying FILETIME.

Note that the TimeSpan is ok to leave without a raw ticks. TimeSpan ctor for ticks just directly stores it, no real cost (unlike DateTime ctor, which does a lot of work).