Embed annotations directly
Annotations need to persist so that Rider can know for example that DefaultDictionary's [] can modify (and not warn about an un-updated collection when using it in the client). But I don't want to publish use of annotations because that would add a dll and dependency for users. So we embed it instead as per citizenmatt's advice.

I don't like doing this because I want to keep the annotations at latest version. But they don't change very often so this is the best compromise for now.